refactor(layout): move shutdown/ to serve/shutdown/ (T124.3.2)
