+++ serve/shutdown/doc.go
